Require SSE4.2 for the driver.
The weakest CPU supported by this driver is Bay Trail-M which supports SSE4.2, and considering that this driver won't support anything older as the iGPUs of the past are much less capable, make a bold choice to refuse to run on ancient hardware.

Leave users a message that their CPU lacks SSE4.2 and likely won't need the driver.

Future patchsets will add AVX2 paths for the VP9 hybrid decode code to speed up the CPU side of things.

v2: Fix compilation errors.